New API models and crash fix

+++ app/src/main/java/com/example/studentass/models/ScheduleDay.kt
@@ -1,9 +1,9 @@
 package com.example.studentass.models
 
 class ScheduleDay(
-    val id : Int,
+    val id : Long,
     val number_day : Int,
-    val group_id : Int,
+    val group_id : Long,
     val numberWeek : Int,
     val coupels : List<ScheduleDayCouple>
 )
